fix menu
