+++ src/layouts/MainLayout/MainLayout.tsx
@@ -1,7 +1,7 @@
 import { FC, PropsWithChildren, useContext, useEffect, useState } from 'react'
 import { IconAlertTriangleFilled } from '@tabler/icons-react'
 import NavigationMenuComponent from '@/components/MainLayout/NavigationMenu'
-import { Link, useLocation, useNavigate } from 'react-router-dom'
+import { useLocation, useNavigate } from 'react-router-dom'
 import { store } from '@/store'
 import { useCart } from '@/store/useCartStoreZustand'
 import { useCustomerData } from '@/store/useCustomerStoreZustand'
@@ -20,8 +20,6 @@ import { useCountDown } from '@/hooks/useCountDown'
 import { UserDropdownMenu } from '@/components/MainLayout/DropDownMenu'
 import { MobileNavigationMenu } from '@/components/MainLayout/MobileMenu'
 import { useCartOrders } from '@/store/useCartStoreOrdersZustand'
-import { SideBarData } from './data/data'
-import { useRolePermissions } from '@/hooks/useRolespermissions'
 
 const MainLayout: FC<PropsWithChildren> = ({ children }) => {
   const [modalState, setModalState] = useState<ModalStateEnum>(ModalStateEnum.off)
@@ -105,35 +103,6 @@ const MainLayout: FC<PropsWithChildren> = ({ children }) => {
         <UserDropdownMenu />
       </nav>
       <div className='w-100 h-screen pt-16 flex bg-background-main'>
-        <div className='w-56 bg-white'>
-          <ul className='list-none p-0 m-0 mt-12 ml-5'>
-            {SideBarData.map((item, index) => {
-              if (item.showInSideBar === false) return null
-              const active = location.pathname === item.path
-              if (item.allowedRoles) {
-                const { hasPermission } = useRolePermissions({ allowedRoles: item.allowedRoles })
-
-                if (!hasPermission) {
-                  return null
-                }
-              }
-              const Icon = item.icon
-              return (
-                <li key={index}>
-                  <Link
-                    to={item.path}
-                    className={`flex items-center mb-7 cursor-pointer gap-3 no-underline ${
-                      active ? 'text-green-1' : 'text-gray-1'
-                    } `}
-                  >
-                    {Icon && <Icon />}
-                    <span className='text-sm'>{item.title}</span>
-                  </Link>
-                </li>
-              )
-            })}
-          </ul>
-        </div>
         <div className='h-full w-full overflow-hidden p-5 flex flex-col gap-4'>
           {!state.user?.is_verified && (
             <div className='w-full z-1 p-5 flex justify-start items-center bg-red-300 text-white rounded-md gap-1'>
