+++ tailwind.config.cjs
@@ -40,6 +40,33 @@ export const theme = {
         foreground: 'hsl(var(--accent-foreground))',
       },
     },
+    borderRadius: {
+      lg: 'var(--radius)',
+      md: 'calc(var(--radius) - 2px)',
+      sm: 'calc(var(--radius) - 4px)',
+    },
+    keyframes: {
+      'accordion-down': {
+        from: {
+          height: '0',
+        },
+        to: {
+          height: 'var(--radix-accordion-content-height)',
+        },
+      },
+      'accordion-up': {
+        from: {
+          height: 'var(--radix-accordion-content-height)',
+        },
+        to: {
+          height: '0',
+        },
+      },
+    },
+    animation: {
+      'accordion-down': 'accordion-down 0.2s ease-out',
+      'accordion-up': 'accordion-up 0.2s ease-out',
+    },
     gridTemplateColumns: {
       13: 'repeat(13, minmax(0, 1fr))',
     },
